Split program into subcommands
I plan to add some more functionality to the program, and dividing it
into discrete subcommands will prove helpful to that end.

I'd like to install the "extra" binaries into libexec, but I can't
figure out how to do that with RubyGems (or if it's even possible),
so into bin they go instead...for now.