Return to original UI
I didn't like the whole subcommand idea; in my opinion, it resulted in
too much typing. This tool is supposed to be _simple_. I kept the basic
structure in place, though, but the command-loading code is less
complicated because the program, as a whole, is less dynamic.

I'll revisit this idea if I ever get around to expanding the power of
USaidWat, but let's keep things simple for the 1.0 release.