Test Cucumber features using included mock server
A basic Sinatra server is included that mocks the Reddit web server,
returning static data instead of "real" dynamic data.

One caveat: Right now, the server has to be run manually before the
Cucumber test suite is run. I'd like to have the server start and
stop automatically, and I put a lot of effort into making that
happen, but I couldn't get it to work properly yet, so this manual
method will have to do for now. However, I'd like to proceed with
testing and don't want to get too hung up on this technicality.