Expose command UI
Instead of controlling output via flags, there are now a series of
subcommands that allow the user to interact with usaidwat in different
ways, each with its own set of flags. The current commands are:

  * tally
  * log

Thor is used to handle subcommands. I'm not entirely a fan of Thor's
help output, nor am I a fan of some of its exit codes, so the use of
this gem is subject to change; however, it does allow me to concretely
set the usaidwat UI, so I can finally release v1.0 of this gem.

Closes #16.