Avoid repeating lvalue when chaining Either monad operations
Since the Right Either passes itself into the block it accepts, we can
always grab its value if need be; no need to repeat the lvalue in the
chain. This is more along the lines of how I intended Either to be used
in the first place; I just didn't think through all the possibilities.