New subcommand: `user`
`usaidwat user` will show an account's creation date and karma scores.

I'm a bit torn on adding this, as I feel it's a bit outside the scope of
usaidwat (I don't want usaidwat to become a general Reddit query tool
per se), but it's useful enough that I think I will go ahead and add it.

Closes #28.