Fix issue when calculating years for a given time delta
The issue arose because a date in Nov. 2007 was marked as "over 8 years
ago" instead of "over 7 years ago". The problem is that years have the
modifier "over" instead of "about", but they were using #round instead
of flooring the float (i.e., using #to_i). #to_i is the correct method
to apply.