Unescape comment bodies before rendering
There was an issue in which comment bodies were not rendered properly in
downterm, because HTML entities were not escaped before passing to
downterm. For example, quoted portions of comments began with "&gt;",
not ">", so downterm did not see them as blockquotes and they were
rendered as paragraphs instead. Escaping before passing to the Markdown
renderer will ensure that marked-up elements are processed correctly.