Avoid installing mercenary 0.3.6
It introduced a small bug with the way help/usage output is printed that
results in parent command flags being duplicated in the help output of
subcommands. I expect this will eventually be fixed, so for now I'll
ensure I skip v0.3.6 (0.3.6 includes no other useful changes for me
anyway).