Monkey patch Snooby to use Reddit's HTTPS URLs
Reddit API calls via Snooby have been failing today, because calls via
HTTP to Reddit's API servers are returning 301 Moved Permanently, which
Snooby appears not to be following. (I don't know if Reddit just started
doing this or not, but it wasn't an issue until now.) I submitted a pull
request to fix this problem in Snooby [1], but until that is accepted, I
will need to monkey patch Snooby so that fix is applied when using
usaidwat. Ugly solution, but it'll work until Snooby can be fixed.

  [1] https://github.com/andkerosine/snooby/pull/2