Replace Snooby with homegrown client
Snooby is old and not updated any more; moreover, it didn't correctly
pin gem versions (such as net-http-persistent), resulting in problems
when updating gems. The public Reddit API is pretty simple and
straightforward, and it'll be easier just to roll my own.

(I'm using Unirest now, which isn't a silver bullet; Unirest is also old
and not really updated anymore. But it should be enough to get by until
I can find a replacement for that, too.)

Fixes #35.