Remove USER_AGENT class constant
This was throwing an error, something about USaidWat::VERSION not being
available, but only once the gem was installed. Not sure why -- probably
has to do with an oddity in how Ruby loads required files that I'm not
aware of. Moving it to the `get` body works fine, so I'll just do that.