Only print out message for ArgumentErrors
At some point, something in Ruby or a dependency changed so that the
_entire_ stack trace was printed when ArgumentErrors were thrown,
instead of just the ArgumentError message. Aside from causing the
Cucumber tests to fail, that was also ugly, as it is not really
necessary for users to see the full stack trace when they failed to call
a subcommand correctly. Thus, USaidWat will now catch ArgumentErrors and
only print out the error message, not the entire stack trace.

Obviously this resulted in some duplicated begin/rescue blocks. I should
probably refactor this code to handle ArgumentErrors in one place, but
that is a bit complicated since the handler has to be contained within a
block, so I will need to write some kind of "springboard" method to
properly handle that.