Word-wrap comment bodies to terminal width
Previously, lines were wrapped by the terminal itself, so sometimes they
wrapped in the middle of words or other awkward spots in the text. Now,
bodies will be wrapped more intelligently, so the line breaks at a word
boundary.

This commit does not pass the Cucumber tests! It's tricky to wrap quoted
Markdown blocks, and that feature is not yet working correctly. I
believe I will have to update the Downterm gem to handle that feature,
which will take some work.